code: move testSummaries stubs 2 levels up
