+++ detox/src/artifacts/ArtifactsManager.test.js
@@ -1,6 +1,6 @@
 const path = require('path');
 const sleep = require('../utils/sleep');
-const testSummaries = require('./templates/plugin/__mocks__/testSummaries.mock');
+const testSummaries = require('./__mocks__/testSummaries.mock');
 const testSuite = require('./templates/plugin/__mocks__/testSuite.mock');
 
 describe('ArtifactsManager', () => {
