+++ detox/src/artifacts/__mocks__/testSummaries.mock.js
