+++ detox/src/artifacts/templates/plugin/StartupAndTestRecorderPlugin.test.js
@@ -1,7 +1,7 @@
 jest.mock('../../../utils/logger.js');
 const StartupAndTestRecorderPlugin = require('./StartupAndTestRecorderPlugin');
 const ArtifactsApi = require('./__mocks__/ArtifactsApi.mock');
-const testSummaries = require('./__mocks__/testSummaries.mock');
+const testSummaries = require('../../__mocks__/testSummaries.mock');
 const ArtifactMock = require('../artifact/__mocks__/ArtifactMock');
 
 describe('StartupAndTestRecorderPlugin', () => {
